Commit 10
Login y Registyro finalizados
Controlar el logueado
Y lista de sitios

+++ Proyectos_de_la_Tesis/Front-End/UnidadFamiliar/src/app/pages/signup/signup.page.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { UserService } from '../../api/user.service';
 import { Router } from '@angular/router';
 import { ToastController } from '@ionic/angular';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { MenuController } from '@ionic/angular';
 
 @Component({
   selector: 'app-signup',
@@ -11,6 +13,7 @@ import { ToastController } from '@ionic/angular';
 export class SignupPage implements OnInit {
 
   userData = { name : '', email : '', telefono : '' , password : '', status : 'Active' };
+  userExiste = {telefono : '' , status : 'Active'};
   responseData: any;
 
   boolean: any;
@@ -19,11 +22,15 @@ export class SignupPage implements OnInit {
   buttonUpdate: any;
   titleSignup: any;
   titleUpdate: any;
+  formularioUsuario: FormGroup;
+  existeUsuario: any;
 
   constructor(
       private authService: UserService,
       private router: Router,
-      private toastController: ToastController
+      private toastController: ToastController,
+      private fb: FormBuilder,
+      public menu: MenuController,
   ) {
     localStorage.getItem('update');
     console.log('update', localStorage.getItem('update') + '==' + null);
@@ -32,15 +39,11 @@ export class SignupPage implements OnInit {
       this.activePassword = false;
       this.buttonRegister = false;
       this.buttonUpdate   = true;
-      this.titleSignup    = false;
-      this.titleUpdate    = true;
     } else if (localStorage.getItem('update') === 'false') {
       this.boolean        = false;
       this.activePassword = true;
       this.buttonRegister = true;
       this.buttonUpdate   = false;
-      this.titleSignup    = true;
-      this.titleUpdate    = false;
     }
   }
 
@@ -53,24 +56,35 @@ export class SignupPage implements OnInit {
   }
 
   ngOnInit() {
+    this.buildForm();
   }
 
   signup() {
     if (this.userData.name && this.userData.email && this.userData.password && this.userData.telefono) {
-      this.authService.postData(JSON.stringify(this.userData), 'signup').then((result) => {
-        this.responseData = result;
-        console.log(this.responseData);
-        if (this.responseData.api_status === 1 && this.responseData.api_http === 200) {
-          localStorage.setItem('userData', JSON.stringify(this.responseData));
-          this.presentToast('User saved successful.');
-          this.router.navigate([ '/menu/login' ]);
+      this.userExiste.telefono = this.userData.telefono;
+      this.userData.telefono = this.userData.telefono;
+      this.authService.postData(JSON.stringify(this.userExiste), 'controlarusuariosrepetidos').then((res) => {
+        this.existeUsuario = res;
+        if (this.existeUsuario.api_status === 1 && this.existeUsuario.api_http === 200) {
+          this.presentToast('El número de teléfono ingresado ya existe.');
+        } else {
+          this.authService.postData(JSON.stringify(this.userData), 'signup').then((result) => {
+            this.responseData = result;
+            console.log(this.responseData);
+            if (this.responseData.api_status === 1 && this.responseData.api_http === 200) {
+              localStorage.setItem('userData', JSON.stringify(this.responseData));
+              this.presentToast('Usuario guardado correctamente.');
+              // this.buildForm();
+              this.router.navigate([ '/menu/login' ]);
+            }
+          }, (err) => {
+            console.log(err);
+            this.presentToast('Falla del servicio.');
+          });
         }
-      }, (err) => {
-        console.log(err);
-        this.presentToast('The service is failed.');
-      });
+       });
     } else {
-      this.presentToast('The data is required.');
+      this.presentToast('Todos campos son necesarios.');
     }
   }
 
@@ -91,66 +105,25 @@ export class SignupPage implements OnInit {
     }
   }
 
-  /*async presentAlert() {
-    const alertController = document.querySelector('ion-alert-controller');
-    await alertController.componentOnReady();
-
-    const alert = await alertController.create({
-      header: 'Alert',
-      subHeader: 'Subtitle',
-      message: 'This is an alert message.',
-      buttons: ['OK']
+  buildForm() {
+    /**
+     * @description Asignamos a la propiedad "formularioUsuario" los campos que se van a controlar de la vista
+     */
+    this.formularioUsuario = this.fb.group({
+      nombre: ['', [Validators.required, Validators.maxLength(30)]],
+      correo: ['', [Validators.required, Validators.email]],
+      numero_contacto: ['', [Validators.required, Validators.minLength(10), Validators.maxLength(10), Validators.pattern(/^[0-9]{5,10}$/)]],
+      password: ['', [Validators.required, Validators.minLength(5), Validators.maxLength(10)]]
     });
-    return await alert.present();
-  }*/
+  }
 
-  async presentAlert() {
-    const alertController = document.querySelector('ion-alert-controller');
-    await alertController.componentOnReady();
+  ionViewWillEnter() {
+    /* this.storage.get('ion_did_tutorial').then(res => {
+      if (res === true) {
+        this.router.navigateByUrl('/app/tabs/schedule');
+      }
+    }); */
 
-    console.log(this.buttonUpdate)  ;
-    if (this.buttonUpdate === false) {
-      const alert = await alertController.create({
-        header: 'Alert',
-        message: 'You are sure the update the information?',
-        buttons: [
-          {
-            text: 'Cancel',
-            role: 'cancel',
-            handler: () => {
-              // this.navCtrl.setRoot(SignupPage);
-            }
-          },
-          {
-            text: 'Yes',
-            handler: () => {
-              this.router.navigate([ '/menu/profile' ]);
-            }
-          }
-        ]
-      });
-      alert.present();
-    } else {
-      const alert = await alertController.create({
-        header: 'Alert',
-        message: 'You are sure the create this account?',
-        buttons: [
-          {
-            text: 'Cancel',
-            role: 'cancel',
-            handler: () => {
-              // this.navCtrl.setRoot(SignupPage);
-            }
-          },
-          {
-            text: 'Yes',
-            handler: () => {
-              this.router.navigate([ '/menu/login' ]);
-            }
-          }
-        ]
-      });
-      alert.present();
-    }
+    this.menu.enable(false);
   }
 }
